Clean for a good diff to review
